api/requirements.txt
aiofiles==23.1.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2024.7.4
chardet==4.0.0
click==7.1.2
dnspython==2.6.1
email-validator==2.0.0
fastapi==0.120.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.16.0
httptools>=0.3.0,<0.7.0
httpx==0.28.1
idna==3.7
itsdangerous==1.1.0
Jinja2==3.1.4
MarkupSafe==2.0.1
orjson==3.9.15
promise==2.3
pydantic>=2
python-dotenv==0.17.1
python-multipart==0.0.18
PyYAML>=5.4.1,<6.1.0
requests==2.32.4
Rx==1.6.1
starlette==0.49.1
typing-extensions==4.13.2
ujson==4.0.2
urllib3==2.6.3
uvicorn==0.13.4
uvloop==0.21.0
watchgod==0.7
websockets==10.0
pymongo==4.16.0
PyJWT==2.10.1